# backend/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WAY-Esports API",
    "version": "1.0.0"
  },
  "servers": [{ "url": "/" }],
  "paths": {
...
    "/api/matches": {
      "get": {
        "summary": "Список матчей",
        "parameters": [
          { "name": "game", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "tournament", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MatchList" } } } } }
      },
      "post": {
        "summary": "Создать матч",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MatchCreate" } } } },
        "responses": { "201": { "description": "Создано" }, "404": { "description": "Tournament not found" } }
      }
    },
    "/api/matches/batch": {
      "post": {
        "summary": "Массовое создание матчей",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/MatchCreate" } } } } },
        "responses": { "201": { "description": "Создано" }, "404": { "description": "Tournament not found" }, "400": { "description": "Ошибка валидации" } }
      }
    },
...
    "/api/matches/{id}": {
      "get": {
        "summary": "Матч по ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Match" } } } }, "404": { "description": "Not found" } }
      }
    },
    "/api/matches/{id}/score": {
      "put": {
        "summary": "Обновить счёт и статистику матча",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MatchScoreUpdate" } } } },
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not found" } }
      }
    },
    "/api/matches/{id}/status": {
      "put": {
        "summary": "Обновить статус матча",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "status": { "type": "string" } }, "required": ["status"] } } } },
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not found" } }
      }
    },
    "/api/teams": {
      "get": {
        "summary": "Список команд",
        "parameters": [
          { "name": "game", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TeamList" } } } } }
      },
      "post": {
        "summary": "Создать команду",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TeamCreate" } } } },
        "responses": { "201": { "description": "Создано" }, "400": { "description": "Имя/тег заняты" } }
      }
    },
    "/api/teams/batch": {
      "post": {
        "summary": "Массовое создание команд",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/TeamCreate" } } } } },
        "responses": { "201": { "description": "Создано" }, "400": { "description": "Имя/тег заняты" } }
      }
    },
    "/api/teams/{id}": {
      "get": {
        "summary": "Команда по ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Team" } } } }, "404": { "description": "Not found" } }
      },
      "put": {
        "summary": "Обновить команду",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TeamUpdate" } } } },
        "responses": { "200": { "description": "OK" }, "403": { "description": "Forbidden" }, "404": { "description": "Not found" } }
      }
    },
    "/api/teams/{id}/members": {
      "post": {
        "summary": "Добавить участника",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "userId": { "type": "string" } }, "required": ["userId"] } } } },
        "responses": { "200": { "description": "OK" }, "400": { "description": "Already member" }, "403": { "description": "Forbidden" }, "404": { "description": "Not found" } }
      }
    },
    "/api/teams/{id}/members/{userId}": {
      "delete": {
        "summary": "Удалить участника",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK" }, "403": { "description": "Forbidden" }, "404": { "description": "Not found" } }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Match": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "game": { "type": "string" },
          "status": { "type": "string" },
          "team1": { "type": "string" },
          "team2": { "type": "string" },
          "tournament": { "type": "string" },
          "startTime": { "type": "string", "format": "date-time" },
          "score": { "type": "object" }
        }
      },
      "MatchList": { "type": "object", "properties": { "success": { "type": "boolean" }, "data": { "type": "array", "items": { "$ref": "#/components/schemas/Match" } } } },
      "MatchCreate": { "type": "object", "properties": { "game": { "type": "string" }, "team1": { "type": "string" }, "team2": { "type": "string" }, "tournament": { "type": "string" }, "startTime": { "type": "string", "format": "date-time" } }, "required": ["game", "team1", "team2", "tournament"] },
      "MatchScoreUpdate": { "type": "object", "properties": { "score": { "type": "object" }, "stats": { "type": "object" }, "winner": { "type": "string" } } },
      "Team": { "type": "object", "properties": { "_id": { "type": "string" }, "name": { "type": "string" }, "tag": { "type": "string" }, "game": { "type": "string" }, "members": { "type": "array", "items": { "type": "string" } } } },
      "TeamList": { "type": "object", "properties": { "success": { "type": "boolean" }, "data": { "type": "array", "items": { "$ref": "#/components/schemas/Team" } } } },
      "TeamCreate": { "type": "object", "properties": { "name": { "type": "string" }, "tag": { "type": "string" }, "game": { "type": "string" } }, "required": ["name", "tag", "game"] },
      "TeamUpdate": { "type": "object" },
...
    }
  }
}